Add flag-based setting for 'erts_de_busy_limit' constant
A prior patch to change the 'erts_de_busy_limit' constant, based on an
environment variable, was rejected by the Ericsson OTP team because
environment variables are no longer used for modifying the VM's
behavior.

Families of flags like "+s" are devoted to the scheduler (except for
"+sss"), but it wasn't obvious to me where a "+"-style flag for
'erts_de_busy_limit' should go.  So, this patch creates a new family
of flags, prefixed with "+z", for miscellaneous VM settings.

This patch creates a new family of flags with the "+z" prefix.  It
further creates a new configuration option called "dbbl" (which is the
first letter of the name dist_buf_busy_limit).  Example usage of this
flag would be "+zdbbl 1048576".

This patch creates an adjustable buffer limit for the amount of data
that may be buffered by the erlang distribution code (in dist.c
specifically).  Before this patch, this hard-coded constant was used:

    #define ERTS_DE_BUSY_LIMIT (128*1024)

When large binaries are transmitted between nodes (or simply a lot of
medium-sized binaries), it is very easy to hit the old 128KB limit.
Processes that use the erlang:system_monitor() BIF to monitor system
events can be spammed by {monitor, busy_dist_port, ...} message tuples
at rates of tens to even hundreds of messages/second.

A larger buffer limit will allow processes to buffer more outgoing
messages over the distribution. When the buffer limit has been
reached, sending processes will be suspended until the buffer size has
shrunk. The buffer limit is per distribution channel. A higher limit
will give lower latency (and higher throughput?) at the expense of
higher memory usage.

A variation of this patch has been in commercial production use in at
least two companies that the author is aware of.  Larger buffer values
can reduce the number of {monitor, busy_dist_port, ...} system
messages drastically, lower overall messaging latencies, and prevent
false timeouts and 'nodedown' messages in extremely busy Mnesia systems.

Test suite: there are two tests:

  a. In erlexec_SUITE.erl to test basic set & get of the value
  b. In distribution_SUITE.erl, to verify that setting +zdbbl very
     low will actually change behavior.
The documentation changes are expanded since the last proposal and
are hopefully much closer to sufficient.